+++ tests/Unit/Framework/ClassResolver/ClassNameFinderTest.php
@@ -16,7 +16,9 @@ final class ClassNameFinderTest extends TestCase
 {
     public function setUp(): void
     {
-        Gacela::bootstrap(__DIR__, static fn (GacelaConfig $config) => $config->setResetCache(true));
+        Gacela::bootstrap(__DIR__, static function (GacelaConfig $config): void {
+            $config->setResetCache(true);
+        });
     }
 
     public function test_no_rules(): void